Update API, use 64 bit ints, fix tests

+++ ticketmatic/settings/communicationanddesign/ordermails/operations.go
@@ -20,7 +20,7 @@ func Getlist(client *ticketmatic.Client, params *ticketmatic.OrderMailTemplateQu
 }
 
 // Get a single order mail template
-func Get(client *ticketmatic.Client, id int) (*ticketmatic.OrderMailTemplate, error) {
+func Get(client *ticketmatic.Client, id int64) (*ticketmatic.OrderMailTemplate, error) {
 	r := client.NewRequest("GET", "/{accountname}/settings/communicationanddesign/ordermails/{id}")
 	r.UrlParameters(map[string]interface{}{
 		"id": id,
@@ -48,7 +48,7 @@ func Create(client *ticketmatic.Client, data *ticketmatic.OrderMailTemplate) (*t
 }
 
 // Modify an existing order mail template
-func Update(client *ticketmatic.Client, id int, data *ticketmatic.OrderMailTemplate) (*ticketmatic.OrderMailTemplate, error) {
+func Update(client *ticketmatic.Client, id int64, data *ticketmatic.OrderMailTemplate) (*ticketmatic.OrderMailTemplate, error) {
 	r := client.NewRequest("PUT", "/{accountname}/settings/communicationanddesign/ordermails/{id}")
 	r.UrlParameters(map[string]interface{}{
 		"id": id,
@@ -71,7 +71,7 @@ func Update(client *ticketmatic.Client, id int, data *ticketmatic.OrderMailTempl
 //
 // Most object types are archivable and can't be deleted: this is needed to ensure
 // consistency of historical data.
-func Delete(client *ticketmatic.Client, id int) error {
+func Delete(client *ticketmatic.Client, id int64) error {
 	r := client.NewRequest("DELETE", "/{accountname}/settings/communicationanddesign/ordermails/{id}")
 	r.UrlParameters(map[string]interface{}{
 		"id": id,
