+++ ticketmatic/settings/ticketsales/locktypes/operations.go
@@ -20,7 +20,7 @@ func Getlist(client *ticketmatic.Client, params *ticketmatic.LockTypeQuery) ([]*
 }
 
 // Get a single lock type
-func Get(client *ticketmatic.Client, id int) (*ticketmatic.LockType, error) {
+func Get(client *ticketmatic.Client, id int64) (*ticketmatic.LockType, error) {
 	r := client.NewRequest("GET", "/{accountname}/settings/ticketsales/locktypes/{id}")
 	r.UrlParameters(map[string]interface{}{
 		"id": id,
@@ -48,7 +48,7 @@ func Create(client *ticketmatic.Client, data *ticketmatic.LockType) (*ticketmati
 }
 
 // Modify an existing lock type
-func Update(client *ticketmatic.Client, id int, data *ticketmatic.LockType) (*ticketmatic.LockType, error) {
+func Update(client *ticketmatic.Client, id int64, data *ticketmatic.LockType) (*ticketmatic.LockType, error) {
 	r := client.NewRequest("PUT", "/{accountname}/settings/ticketsales/locktypes/{id}")
 	r.UrlParameters(map[string]interface{}{
 		"id": id,
@@ -71,7 +71,7 @@ func Update(client *ticketmatic.Client, id int, data *ticketmatic.LockType) (*ti
 //
 // Most object types are archivable and can't be deleted: this is needed to ensure
 // consistency of historical data.
-func Delete(client *ticketmatic.Client, id int) error {
+func Delete(client *ticketmatic.Client, id int64) error {
 	r := client.NewRequest("DELETE", "/{accountname}/settings/ticketsales/locktypes/{id}")
 	r.UrlParameters(map[string]interface{}{
 		"id": id,
